+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -79,6 +79,7 @@ public let masterRuleList = RuleList(rules: [
     OperatorFunctionWhitespaceRule.self,
     OperatorUsageWhitespaceRule.self,
     OverriddenSuperCallRule.self,
+    OverrideInExtensionRule.self,
     PatternMatchingKeywordsRule.self,
     PrivateOutletRule.self,
     PrivateOverFilePrivateRule.self,